переделана фильтрация
